Fix port reference for label editing
The wrong port variable was being used when attempting to add the
base_url:port as a label in the metric collection. This has now been
fixed and tested.

Version bumped to 0.18.2.

Added some info in the readme on whats different in the bw specific,
how to compile and how to debug.